app/requirements.txt
# Flask==2.2.5
# psycopg2-binary==2.9.7
# gunicorn==20.1.0
Flask==2.3.2
psycopg2-binary==2.9.9
gunicorn==20.1.0
